+++ tableaudocumentapi/datasource.py
@@ -1,11 +1,5 @@
-###############################################################################
-#
-# Datasource - A class for writing datasources to Tableau files
-#
-###############################################################################
 import collections
 import itertools
-import random
 import xml.etree.ElementTree as ET
 import xml.sax.saxutils as sax
 from uuid import uuid4
@@ -17,7 +11,7 @@ from tableaudocumentapi.xfile import xml_open
 
 ########
 # This is needed in order to determine if something is a string or not.  It is necessary because
-# of differences between python2 (basestring) and python3 (str).  If python2 support is every
+# of differences between python2 (basestring) and python3 (str).  If python2 support is ever
 # dropped, remove this and change the basestring references below to str
 try:
     basestring
@@ -36,7 +30,7 @@ def _get_metadata_xml_for_field(root_xml, field_name):
 
 
 def _is_used_by_worksheet(names, field):
-    return any((y for y in names if y in field.worksheets))
+    return any(y for y in names if y in field.worksheets)
 
 
 class FieldDictionary(MultiLookupDict):
@@ -88,25 +82,32 @@ def base36encode(number):
     return sign + base36
 
 
-def make_unique_name(dbclass):
+def _make_unique_name(dbclass):
     rand_part = base36encode(uuid4().int)
     name = dbclass + '.' + rand_part
     return name
 
 
 class ConnectionParser(object):
+    """Parser for detecting and extracting connections from differing Tableau file formats."""
 
     def __init__(self, datasource_xml, version):
         self._dsxml = datasource_xml
         self._dsversion = version
 
     def _extract_federated_connections(self):
-        return list(map(Connection, self._dsxml.findall('.//named-connections/named-connection/*')))
+        connections = list(map(Connection, self._dsxml.findall('.//named-connections/named-connection/*')))
+        # 'sqlproxy' connections (Tableau Server Connections) are not embedded into named-connection elements
+        # extract them manually for now
+        connections.extend(map(Connection, self._dsxml.findall("./connection[@class='sqlproxy']")))
+        return connections
 
     def _extract_legacy_connection(self):
         return list(map(Connection, self._dsxml.findall('connection')))
 
     def get_connections(self):
+        """Find and return all connections based on file format version."""
+
         if float(self._dsversion) < 10:
             connections = self._extract_legacy_connection()
         else:
@@ -135,16 +136,11 @@ class FilterParser(object):
 
 
 class Datasource(object):
-    """
-    A class for writing datasources to Tableau files.
+    """A class representing Tableau Data Sources, embedded in workbook files or
+    in TDS files.
 
     """
 
-    ###########################################################################
-    #
-    # Public API.
-    #
-    ###########################################################################
     def __init__(self, dsxml, filename=None):
         """
         Constructor.  Default is to create datasource from xml.
@@ -156,6 +152,7 @@ class Datasource(object):
         self._name = self._datasourceXML.get('name') or self._datasourceXML.get(
             'formatted-name')  # TDS files don't have a name attribute
         self._version = self._datasourceXML.get('version')
+        self._caption = self._datasourceXML.get('caption', '')
         self._connection_parser = ConnectionParser(
             self._datasourceXML, version=self._version)
         self._connections = self._connection_parser.get_connections()
@@ -172,13 +169,15 @@ class Datasource(object):
 
     @classmethod
     def from_file(cls, filename):
-        """Initialize datasource from file (.tds)"""
+        """Initialize datasource from file (.tds ot .tdsx)"""
 
-        dsxml = xml_open(filename, cls.__name__.lower()).getroot()
+        dsxml = xml_open(filename, 'datasource').getroot()
         return cls(dsxml, filename)
 
     @classmethod
     def from_connections(cls, caption, connections):
+        """Create a new Data Source give a list of Connections."""
+
         root = ET.Element('datasource', caption=caption, version='10.0', inline='true')
         outer_connection = ET.SubElement(root, 'connection')
         outer_connection.set('class', 'federated')
@@ -186,7 +185,7 @@ class Datasource(object):
         for conn in connections:
             nc = ET.SubElement(named_conns,
                                'named-connection',
-                               name=make_unique_name(conn.dbclass),
+                               name=_make_unique_name(conn.dbclass),
                                caption=conn.server)
             nc.append(conn._connectionXML)
         return cls(root)
@@ -221,23 +220,28 @@ class Datasource(object):
 
         xfile._save_file(self._filename, self._datasourceTree, new_filename)
 
-    ###########
-    # name
-    ###########
     @property
     def name(self):
         return self._name
 
-    ###########
-    # version
-    ###########
     @property
     def version(self):
         return self._version
 
-    ###########
-    # connections
-    ###########
+    @property
+    def caption(self):
+        return self._caption
+
+    @caption.setter
+    def caption(self, value):
+        self._datasourceXML.set('caption', value)
+        self._caption = value
+
+    @caption.deleter
+    def caption(self):
+        del self._datasourceXML.attrib['caption']
+        self._caption = ''
+
     @property
     def connections(self):
         return self._connections
@@ -252,6 +256,11 @@ class Datasource(object):
     ###########
     # fields
     ###########
+    def clear_repository_location(self):
+        tag = self._datasourceXML.find('./repository-location')
+        if tag is not None:
+            self._datasourceXML.remove(tag)
+
     @property
     def fields(self):
         if not self._fields:
@@ -259,6 +268,8 @@ class Datasource(object):
         return self._fields
 
     def _get_all_fields(self):
+        # Some columns are represented by `column` tags and others as `metadata-record` tags
+        # Find them all and chain them into one dictionary
         column_field_objects = self._get_column_objects()
         existing_column_fields = [x.id for x in column_field_objects]
         metadata_only_field_objects = (x for x in self._get_metadata_objects() if x.id not in existing_column_fields)