+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 06 (11 January 2017)
+
+* Initial SQL and query banding support (#123)
+* Fixed bug in xfiles to allow opening workbooks with external file caches (#117, #118)
+* Code Cleanup (#120, #121)
+* Added Py36 support (#124)
+* Switched to pycodestyle from pip8 on travis runs (#124)
+
+## 05 (01 November 2016)
+
+* Added ability to set the port for connections (#97)
+* Added ability to read and write caption for datasources (#99)
+* Added documentation
+
+## 0.4 (07 October 2016)
+
+* Add ability to remove repository location (#86)
+* Fixed bug in connection parsing when federated connections are present (#87)
+* Fixed bug in UNICODE support (#80)
+
 ## 0.3 (31 August 2016)
 
 * Added basic connection class retargeting (#65)